% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calctime}
\alias{calctime}
\title{Calculate the time needed}
\usage{
calctime(hyp, n.items)
}
\arguments{
\item{hyp}{Hypothesis from setup.hypothesis function}

\item{n.items}{Number of items}
}
\value{

}
\description{
Calculate the time needed
}
