% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ncps.R
\name{calculate_ncps}
\alias{calculate_ncps}
\title{Calculate noncentrality parameters}
\usage{
calculate_ncps(
  hyp,
  stat = c("Wald", "LR", "Score", "Gradient"),
  n = 1,
  sampling = FALSE,
  sampling.npers = 10^4,
  approx.npers = 10^4,
  SE.type = "Oakes",
  sampling.mat = "ApproxFisher"
)
}
\arguments{
\item{hyp}{Hypothesis Object created by the setup_hypothesis function}

\item{stat}{character vector containing the statistics to be calculated. Options are "Wald","LR","Score", and "Gradient". All statistics are included by default.}

\item{n}{integer, number of persons that the noncentrality parameters refer to (default is 1)}

\item{sampling}{boolean, TRUE gives the sampling-based estimate instead of the analytical, recommended for larger itemsets}

\item{sampling.npers}{integer, sample size for the sampling-based approach}

\item{approx.npers}{integer, sample size for approximating the Fisher expected information matrix in the sampling-based approach}

\item{SE.type}{Method for calculation of the observed information matrix used for calculating the statistics in the sampling-based approach ("Oakes" by default)}

\item{sampling.mat}{Approach to calculate the information matrix used for calculating the statistics in the sampling-based approach. By default ("ApproxFisher"), an sampling-based approximation of the expected Fisher matrix is calculated using an observed information matrix of the type SE.type}
}
\value{
numerical vector containing the noncentrality parameters
}
\description{
#' Wrapper function for the different analytical and sampling-based functions
}
\examples{
dat <- expand.table(LSAT7)
mirtfit <- mirt(dat,1,verbose = FALSE)
hyp <- setup_hypothesis(type = "1PLvs2PL", altpars = mirtfit)
ncps <- calculate_ncps(hyp=hyp)

}
