% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{coef.long}
\alias{coef.long}
\title{extract coefs from mirtfit or element created by shortcoef}
\usage{
\method{coef}{long}(mirtfit = NULL, itemtype = NULL, shortcoef = NULL)
}
\arguments{
\item{mirtfit}{object created from mirt}

\item{itemtype}{optional, itemtype as string}

\item{shortcoef}{alternative, object created from shortcoef}
}
\value{

}
\description{
extract coefs from mirtfit or element created by shortcoef
}
