% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{coef_short}
\alias{coef_short}
\title{extract coefs from mirtfit}
\usage{
coef_short(mirtfit, itemtype = NULL)
}
\arguments{
\item{mirtfit}{object created from mirt()}

\item{itemtype}{optional, itemtype as string}
}
\value{

}
\description{
extract coefs from mirtfit
}
\examples{

dat <- expand.table(LSAT7)
mirtfit <- mirt(dat,1)
pars = coef_short(mirtfit)

}
