% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_ncp}
\alias{get_ncp}
\title{Extract ncp from vector of chi² distributed values}
\usage{
get_ncp(chii, df)
}
\arguments{
\item{chii}{numeric vector}

\item{df}{integer, degrees of freedom}
}
\value{

}
\description{
Used in the sampling based ncps
}
