% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml.fit.R
\name{grad_obs}
\alias{grad_obs}
\title{Calculate Gradient statistic}
\usage{
grad_obs(fitted, lx = NULL)
}
\arguments{
\item{fitted}{Object created from mml.fit}
}
\value{

}
\description{
Calculate Gradient statistic
}
\examples{

altpars <- list(
a = rlnorm(5,sdlog = .4),
d = rnorm(5))
hyp <- setup_hypothesis(type = "1PLvs2PL", altpars = altpars)
data <- setup.data(hyp=hyp,n=500)
fitted <- mml.fit(data = data,hyp = hyp)
grad_obs(fitted)

}
