% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infmat.R
\name{infmat}
\alias{infmat}
\title{Calculate information matrix (expected / observed)}
\usage{
infmat(pars, method = "mirtFisher", simbased.npers = 2e+05, data = NULL)
}
\arguments{
\item{pars}{Parameter Set}

\item{method}{string, method to be used, e.g. "mirtFisher" for the Fisher-expected information matrix calculated using the mirt package}

\item{simbased.npers}{integer, number of persons in the sampling based approach}

\item{data}{dataset for observed matrix methods}
}
\value{

}
\description{
Calculate information matrix (expected / observed)
}
