% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.functions.R
\name{load.functions}
\alias{load.functions}
\title{Load IRT model functions and derivatives}
\usage{
load.functions(model, multi = FALSE)
}
\arguments{
\item{model}{Desired Model (2PL,3PL, GPCM)}

\item{multi}{Boolean - Multidimensional? (available for 2PL)}
}
\value{

}
\description{
Load IRT model functions and derivatives
}
