% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml.fit.R
\name{mml.fit}
\alias{mml.fit}
\title{Fit restricted and unrestriced model by ML using mirt}
\usage{
mml.fit(hyp, data, infmat.method = "Oakes", free_mean = FALSE)
}
\arguments{
\item{hyp}{Hypothesis Object created from the setup_hypothesis function}

\item{data}{dataframe with data to be fitted}

\item{infmat.method}{Method used for calculating the infmat used for the variance-covariance matrix, e.g. "Oakes" (default)}
}
\value{

}
\description{
Fit restricted and unrestriced model by ML using mirt
}
\examples{

altpars <- list(
a = rlnorm(5,sdlog = .4),
d = rnorm(5))
hyp <- setup_hypothesis(type = "1PLvs2PL", altpars = altpars)
data <- setup.data(hyp=hyp,n=500)
fitted <- mml.fit(data = data,hyp = hyp)

}
