% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ncps.R
\name{ncp.lr}
\alias{ncp.lr}
\title{Analytical ncp for the LR statistic}
\usage{
ncp.lr(hyp, n = 1, parsr = NULL)
}
\arguments{
\item{hyp}{Hypothesis Object created from the setup_hypothesis function}

\item{n}{integer, number of persons that the ncp refers to (default 1)}

\item{parsr}{Restricted parameters}
}
\value{

}
\description{
Analytical ncp for the LR statistic
}
