% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ncps.R
\name{ncp.sim}
\alias{ncp.sim}
\title{Calculate sampling based ncps (Wrapper)}
\usage{
ncp.sim(hyp, stat = "all", n = 1, n.pers = NULL, runs = 3)
}
\arguments{
\item{hyp}{Hypothesis Object created from the setup_hypothesis function}

\item{stat}{character vector containing the statistics to be calculated. Options are "Wald","LR","Score". Also, "all" (default) for all three.}

\item{n}{integer, number of persons that the ncp refers to (default 1)}

\item{n.pers}{integer, Number of persons in the artificial datasets}

\item{runs}{integer, Number of artificial datasets}
}
\value{
numerical vector containing the ncps
}
\description{
Calculate sampling based ncps (Wrapper)
}
\examples{
dat <- expand.table(LSAT7)
mirtfit <- mirt(dat,1)
pars = coef_short(mirtfit)
hyp <- setup_hypothesis(type = "1PLvs2PL", altpars = pars)
ncps <- ncp.sim(hyp=hyp)
}
