% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ncps.R
\name{ncp.sim.wald}
\alias{ncp.sim.wald}
\title{Sampling-based ncp for the Wald statistic}
\usage{
ncp.sim.wald(hyp, n = 1, simbased.npers)
}
\arguments{
\item{hyp}{Hypothesis Object created from the setup_hypothesis function}

\item{n}{integer, number of persons that the ncp refers to (default 1)}

\item{simbased.npers}{integer, Number of persons used in the sampling based approach}
}
\value{

}
\description{
Uses a shortcut not available for LR and score (see paper)
}
