% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ncps.R
\name{ncp.wald}
\alias{ncp.wald}
\title{Analytical ncp for the Wald statistic}
\usage{
ncp.wald(hyp, n = 1)
}
\arguments{
\item{hyp}{Hypothesis Object created from the setup_hypothesis function}

\item{n}{integer, number of persons that the ncp refers to (default 1)}
}
\value{

}
\description{
Analytical ncp for the Wald statistic
}
