% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{power}
\alias{power}
\title{Calculate power}
\usage{
power(hyp, ncp, ssize, alpha = 0.05)
}
\arguments{
\item{hyp}{Hypothesis Object created from the setup_hypothesis function}

\item{ncp}{numeric, Noncentrality parameter for n=1}

\item{ssize}{interger, sample size}

\item{alpha}{numeric, alpha niveau, e.g. .05 (default)}
}
\value{

}
\description{
Calculate power
}
\examples{

dat <- expand.table(LSAT7)
mirtfit <- mirt(dat,1,verbose = FALSE)
hyp <- setup_hypothesis(type = "1PLvs2PL", altpars = mirtfit)
ncps <- calculate_ncps(hyp=hyp)
power(hyp=hyp,ncp=ncps,alpha=.05,ssize=500)

}
