% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.irtpwrresult.R
\name{print.summary.irtpwrresult}
\alias{print.summary.irtpwrresult}
\title{Print Summary of the search result}
\usage{
\method{print}{summary.irtpwrresult}(x, ...)
}
\arguments{
\item{x}{Object of class irtpwrresult as created by the irtpwr function}

\item{...}{additional arguments to be passed}
}
\value{
An object of class summary.irtpwrresult
}
\description{
Print Summary of the search result
}
\examples{
#Load a simulation function
simfun = example.simfun('ttest')
# Perform the search
ds = find.design(simfun = simfun, boundaries = c(100,300), power = .95)
# Output the results
summary(ds)
}
