% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwrml-package.R
\docType{package}
\name{pwrml-package}
\alias{pwrml}
\alias{pwrml-package}
\title{pwrml: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
