% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{save.error}
\alias{save.error}
\title{Helper function for simulation study (used in evaluation_dist.R)}
\usage{
save.error(e, x)
}
\arguments{
\item{e}{Error}

\item{x}{Additional}
}
\value{

}
\description{
Helper function for simulation study (used in evaluation_dist.R)
}
