% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml.fit.R
\name{score_obs}
\alias{score_obs}
\title{Calculate Score statistic}
\usage{
score_obs(fitted)
}
\arguments{
\item{fitted}{Object created from mml.fit}
}
\value{

}
\description{
Calculate Score statistic
}
\examples{

altpars <- list(
a = rlnorm(5,sdlog = .4),
d = rnorm(5))
hyp <- setup_hypothesis(type = "1PLvs2PL", altpars = altpars)
data <- setup.data(hyp=hyp,n=500)
fitted <- mml.fit(data = data,hyp = hyp)
score_obs(fitted)

}
