% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.data.R
\name{setup.data}
\alias{setup.data}
\title{Create artificial dataset from alternative hypothesis}
\usage{
setup.data(hyp, n, dist.fun = rnorm)
}
\arguments{
\item{hyp}{Hypothesis Object created from the setup_hypothesis function}

\item{n}{integer, number of persons}

\item{dist.fun}{function to generate the person parameters}
}
\value{

}
\description{
Create artificial dataset from alternative hypothesis
}
\examples{

altpars <- list(
a = rlnorm(5,sdlog = .4),
d = rnorm(5))
hyp <- setup_hypothesis(type = "1PLvs2PL", altpars = altpars)
data <- setup.data(hyp=hyp,n=500)

}
