% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ssize}
\alias{ssize}
\title{Calculate sample size}
\usage{
ssize(hyp, ncp, power = 0.8, alpha = 0.05)
}
\arguments{
\item{hyp}{Hypothesis Object created from the setup_hypothesis function}

\item{ncp}{numeric, Noncentrality parameter for n=1}

\item{power}{numeric, desired power, e.g. .8 (default)}

\item{alpha}{numeric, alpha niveau, e.g. .05 (default)}
}
\value{

}
\description{
Calculate sample size
}
\examples{

dat <- expand.table(LSAT7)
mirtfit <- mirt(dat,1,verbose = FALSE)
hyp <- setup_hypothesis(type = "1PLvs2PL", altpars = mirtfit)
ncps <- calculate_ncps(hyp=hyp)
ssize(hyp=hyp,ncp=ncps,alpha=.05,power=.80)

}
