% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml.fit.R
\name{stat_obs}
\alias{stat_obs}
\title{Calculate statistics from fitted mirt models}
\usage{
stat_obs(fitted, stat = c("Wald", "LR", "Score", "Gradient"))
}
\arguments{
\item{fitted}{Object created by mml.fit function.}
}
\value{

}
\description{
Calculate statistics from fitted mirt models
}
