% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ncps.R
\name{calculate_ncps}
\alias{calculate_ncps}
\title{Calculate ncps (analytical/sampling based, Wald/LR/Score)}
\usage{
calculate_ncps(
  hyp,
  stat = c("Wald", "LR", "Score", "Gradient"),
  n = 1,
  sampling = FALSE,
  sampling.npers = 10^4,
  approx.npers = 10^4
)
}
\arguments{
\item{hyp}{Hypothesis Object created from the setup_hypothesis function}

\item{stat}{character vector containing the statistics to be calculated. Options are "Wald","LR","Score". Also, "all" (default) for all three.}

\item{n}{integer, number of persons that the ncp refers to (default 1)}

\item{sampling}{boolean, TRUE gives the simulation based statistics instead of the analytical}

\item{n.pers}{passed to the ncp.sim function}
}
\value{
numerical vector containing the ncps
}
\description{
Wrapper function for the different ncp functions
}
\examples{
dat <- expand.table(LSAT7)
mirtfit <- mirt(dat,1)
pars = coef_short(mirtfit)
hyp <- setup_hypothesis(type = "1PLvs2PL", altpars = pars)
ncps <- calculate_ncps(hyp=hyp)

}
