% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infmat.R
\name{infmat}
\alias{infmat}
\title{Calculate information matrix (expected / observed)}
\usage{
infmat(
  pars,
  method = "Fisher",
  approx.npers = 10^6,
  data = NULL,
  multigroup = FALSE,
  model = 1,
  itemtype = "2PL",
  NCYCLES = 5000,
  SE.type = "Oakes"
)
}
\arguments{
\item{pars}{Parameter Set}

\item{method}{string, method to be used, e.g. "Fisher" for the Fisher-expected information matrix calculated using the mirt package or "ApproxFisher" for its approximation}

\item{approx.npers}{integer, number of persons in the sampling based approach}

\item{data}{dataset for observed matrix methods}

\item{multigroup}{Boolean, multigroup model?}

\item{model}{Specific mirt model}

\item{itemtype}{String, itemtype}
}
\value{

}
\description{
Calculate information matrix (expected / observed)
}
