% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.functions.R
\name{load.functions}
\alias{load.functions}
\title{Load IRT model functions and derivatives}
\usage{
load.functions(model, multi = FALSE)
}
\arguments{
\item{model}{character, Desired Model (2PL,3PL, GPCM)}

\item{multi}{logical, multidimensional model if TRUE (available for 2PL)}
}
\value{
nothing
}
\description{
This is a helper function used to generate custom hypotheses. See the 'adding_hypotheses' vignette.
}
\examples{

funs = load.functions('2PL')

}
