% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml.fit.R
\name{mml.fit}
\alias{mml.fit}
\title{Fit restricted and unrestriced model by ML using mirt}
\usage{
mml.fit(
  hyp,
  data,
  infmat.unres = "Fisher",
  infmat.res = "Fisher",
  free_mean = FALSE,
  approx.npers = 10^6,
  NCYCLES = 5000,
  SE.type = "Oakes"
)
}
\arguments{
\item{hyp}{Hypothesis Object created from the setup_hypothesis function}

\item{data}{dataframe with data to be fitted}

\item{infmat.unres}{String, type of the information matrix for the unrestricted model, "Fisher" or "ApproxFisher" are currently implemented}

\item{infmat.res}{String, type of the information matrix for the restricted model, "Fisher" or "ApproxFisher" are currently implemented}

\item{free_mean}{boolean, option to estimate free means between groups}

\item{approx.npers}{integer, sample size for approximating the Fisher expected information matrix}
}
\value{

}
\description{
Fit restricted and unrestriced model by ML using mirt
}
\examples{

altpars <- list(
a = rlnorm(5,sdlog = .4),
d = rnorm(5))
hyp <- setup_hypothesis(type = "1PLvs2PL", altpars = altpars)
data <- setup.data(hyp=hyp,n=500)
fitted <- mml.fit(data = data,hyp = hyp)

}
