% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ncps.R
\name{ncp.score}
\alias{ncp.score}
\title{Analytical noncentrality parameter for the Score statistic}
\usage{
ncp.score(hyp, n = 1, parsr = NULL)
}
\arguments{
\item{hyp}{Hypothesis Object created by the setup_hypothesis function}

\item{n}{integer, number of persons that the noncentrality parameters refer to (default is 1)}

\item{parsr}{Restricted parameters}
}
\value{

}
\description{
Analytical noncentrality parameter for the Score statistic
}
