% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ncps.R
\name{ncp.wald}
\alias{ncp.wald}
\title{Analytical noncentrality parameter for the Wald statistic}
\usage{
ncp.wald(hyp, n = 1)
}
\arguments{
\item{hyp}{Hypothesis Object created by the setup_hypothesis function}

\item{n}{integer, number of persons that the noncentrality parameters refer to (default is 1)}
}
\value{

}
\description{
Analytical noncentrality parameter for the Wald statistic
}
