% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pars.long}
\alias{pars.long}
\title{extract coefs from mirtfit or element created by shortcoef}
\usage{
pars.long(pars, itemtype, from.mirt = FALSE)
}
\arguments{
\item{itemtype}{optional, itemtype as string}

\item{mirtfit}{object created from mirt}

\item{shortcoef}{alternative, object created from shortcoef}
}
\value{

}
\description{
extract coefs from mirtfit or element created by shortcoef
}
