% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.irtpwrresult.R
\name{print.summary.irtpwrresult}
\alias{print.summary.irtpwrresult}
\title{Print Summary of the search result}
\usage{
\method{print}{summary.irtpwrresult}(x, ...)
}
\arguments{
\item{x}{Object of class irtpwrresult as created by the irtpwr function}

\item{...}{additional arguments to be passed}
}
\value{
An object of class summary.irtpwrresult
}
\description{
Print Summary of the search result
}
\examples{
\donttest{

library(mirt)
dat <- expand.table(LSAT7)
mirtfit <- mirt(dat,1,verbose = FALSE)
hyp <- setup.hypothesis(type = '1PLvs2PL', altpars = mirtfit)
res <- irtpwr(hyp=hyp,alpha=.05,power =.8)
summary(res)
}
}
