% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtpwr-package.R
\docType{package}
\name{irtpwr-package}
\alias{irtpwr}
\alias{irtpwr-package}
\title{irtpwr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
