% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.irtpwrresult.R
\name{summary.irtpwrresult}
\alias{summary.irtpwrresult}
\title{Summary of the Power Analysis}
\usage{
\method{summary}{irtpwrresult}(object, ..., power = NULL, N = NULL, alpha = NULL)
}
\arguments{
\item{object}{Object of class irtpwrresult as created by the irtpwr function}

\item{...}{additional arguments to be passed}
}
\value{
An object of class summary.irtpwrresult
}
\description{
Output the resulting power or sample size
}
\examples{
#Load a simulation function
simfun = example.simfun('ttest')
# Perform the search
ds = find.design(simfun = simfun, boundaries = c(100,300), power = .95)
# Output the results
summary(ds)
}
